requirements_cuda10.txt
--extra-index-url https://developer.download.nvidia.com/compute/redist
--find-links https://download.pytorch.org/whl/torch_stable.html

compressai==1.1.8
DLLogger @ git+https://github.com/NVIDIA/dllogger.git@26a0f8f1958de2c0c460925ff6102a4d2486d6cc
numpy==1.21.0
nvidia-dali-cuda102==1.5.0
pandas==1.3.0
Pillow==8.3.1
psutil==5.8.0
torch==1.9.1+cu102
torchvision==0.10.1+cu102
